+++ lib/rubocop/cop/rake/class_definition_in_task.rb
@@ -3,7 +3,7 @@
 module RuboCop
   module Cop
     module Rake
-      # This cop detects method definition in a task or namespace,
+      # This cop detects class or module definition in a task or namespace,
       # because it is defined to the top level.
       # It is confusing because the scope looks in the task or namespace,
       # but actually it is defined to the top level.
@@ -11,36 +11,38 @@ module RuboCop
       # @example
       #   # bad
       #   task :foo do
-      #     def helper_method
-      #       do_something
+      #     class C
       #     end
       #   end
       #
       #   # bad
       #   namespace :foo do
-      #     def helper_method
-      #       do_something
+      #     module M
       #     end
       #   end
       #
       #   # good - It is also defined to the top level,
       #   #        but it looks expected behavior.
-      #   def helper_method
+      #   class C
       #   end
       #   task :foo do
       #   end
       #
-      class MethodDefinitionInTask < Cop
-        MSG = 'Do not define a method in rake task, because it will be defined to the top level.'
+      class ClassDefinitionInTask < Cop
+        MSG = 'Do not define a %<type>s in rake task, because it will be defined to the top level.'
 
-        def on_def(node)
+        def on_class(node)
           return if Helper::ClassDefinition.in_class_definition?(node)
           return unless Helper::TaskDefinition.in_task_or_namespace?(node)
 
           add_offense(node)
         end
 
-        alias on_defs on_def
+        def message(node)
+          format(MSG, type: node.type)
+        end
+
+        alias on_module on_class
       end
     end
   end